Logger Service Tests Minor Refactoring

+++ candle-api/src/main/java/com/candle/api/model/HistoryResponse.java
@@ -4,6 +4,7 @@ import com.candle.common.model.Candle;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -13,35 +14,36 @@ import java.util.List;
  * not an array of objects — this avoids repeated field-name overhead and
  * matches the TradingView Lightweight Charts / UDF contract exactly.
  *
+ * BigDecimal is used for price arrays to preserve the exact precision
+ * stored in NUMERIC(19,8) database columns. Jackson serialises BigDecimal
+ * as a plain JSON number without scientific notation by default.
+ *
  * <pre>
  * {
  *   "s": "ok",
  *   "t": [1620000000, 1620000060, ...],   // unix seconds
- *   "o": [29500.5,    29501.0,    ...],   // open
- *   "h": [29510.0,    29505.0,    ...],   // high
- *   "l": [29490.0,    29500.0,    ...],   // low
- *   "c": [29505.0,    29502.0,    ...],   // close
- *   "v": [10,         8,          ...]    // volume (tick count)
+ *   "o": [29500.50000000, 29501.00, ...], // open  (BigDecimal)
+ *   "h": [29510.00,       29505.00, ...], // high
+ *   "l": [29490.00,       29500.00, ...], // low
+ *   "c": [29505.00,       29502.00, ...], // close
+ *   "v": [10,             8,        ...]  // volume (tick count)
  * }
  * </pre>
  *
- * When there are no candles in the requested range:
- * <pre>{ "s": "no_data" }</pre>
- *
- * On a validation error:
- * <pre>{ "s": "error", "errmsg": "Unknown interval: 2x. Supported: ..." }</pre>
+ * When there are no candles: <pre>{ "s": "no_data" }</pre>
+ * On a validation error:     <pre>{ "s": "error", "errmsg": "..." }</pre>
  */
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record HistoryResponse(
 
-        @JsonProperty("s")      String   s,
-        @JsonProperty("t")      long[]   t,
-        @JsonProperty("o")      double[] o,
-        @JsonProperty("h")      double[] h,
-        @JsonProperty("l")      double[] l,
-        @JsonProperty("c")      double[] c,
-        @JsonProperty("v")      long[]   v,
-        @JsonProperty("errmsg") String   errmsg
+        @JsonProperty("s")      String       s,
+        @JsonProperty("t")      long[]       t,
+        @JsonProperty("o")      BigDecimal[] o,
+        @JsonProperty("h")      BigDecimal[] h,
+        @JsonProperty("l")      BigDecimal[] l,
+        @JsonProperty("c")      BigDecimal[] c,
+        @JsonProperty("v")      long[]       v,
+        @JsonProperty("errmsg") String       errmsg
 
 ) {
 
@@ -57,12 +59,12 @@ public record HistoryResponse(
         }
 
         int n = candles.size();
-        long[]   t = new long[n];
-        double[] o = new double[n];
-        double[] h = new double[n];
-        double[] l = new double[n];
-        double[] c = new double[n];
-        long[]   v = new long[n];
+        long[]       t = new long[n];
+        BigDecimal[] o = new BigDecimal[n];
+        BigDecimal[] h = new BigDecimal[n];
+        BigDecimal[] l = new BigDecimal[n];
+        BigDecimal[] c = new BigDecimal[n];
+        long[]       v = new long[n];
 
         for (int i = 0; i < n; i++) {
             Candle candle = candles.get(i);
@@ -77,18 +79,12 @@ public record HistoryResponse(
         return new HistoryResponse("ok", t, o, h, l, c, v, null);
     }
 
-    /**
-     * No data available for the requested symbol/interval/range.
-     * TradingView renders a gap on the chart when it receives this.
-     */
+    /** No data available for the requested symbol/interval/range. */
     public static HistoryResponse noData() {
         return new HistoryResponse("no_data", null, null, null, null, null, null, null);
     }
 
-    /**
-     * Validation or server error.
-     * @param message human-readable description returned in "errmsg"
-     */
+    /** Validation or server error. */
     public static HistoryResponse error(String message) {
         return new HistoryResponse("error", null, null, null, null, null, null, message);
     }
